requirements.txt
APScheduler==3.10.4
PyYAML==6.0.1
apprise==1.7.4
croniter==2.0.1

